gc_speech_bridge/requirements.txt
PyAudio
google-cloud-speech
six==1.10.0
google-api-python-client
pyyaml
protobuf-to-dict
